Change kernel insert order (#1420748)
The desired behavior for boot entry lists is to have the non-debug
version of a kernel appear before the debug version of the kernel.

This behavior was originally swapped around to accomodate an issue where
the debug arguments were being populated in all entries. This original
issue has been resolved and the insertion order needs to be swapped
back to the desired behavior.

Resolves: rhbz#1420748